+++ examples/exahype2/ccz4/Properties.f90
@@ -9,7 +9,7 @@ RECURSIVE SUBROUTINE InitParameters(STRLEN,PARSETUP)
 	IMPLICIT NONE  
 	integer          :: STRLEN
 	character(len=STRLEN) :: PARSETUP
-	real :: igamma
+	REAL(8) :: igamma
 
 	ICType=trim(parsetup)
 	print *, "****************************************************************"
@@ -19,37 +19,37 @@ RECURSIVE SUBROUTINE InitParameters(STRLEN,PARSETUP)
         !ICType   = 'CCZ4MinkowskiSrc'
 	EQN%Pi    = ACOS(-1.0)
 #if defined(CCZ4GRMHD) || defined(GRMHD)
-    EQN%DivCleaning_a=1.0
+    EQN%DivCleaning_a=1.0D0
 #else
-    EQN%DivCleaning_a=0.0
+    EQN%DivCleaning_a=0.0D0
 #endif
     !
 	select case(ICType)
         case('GaugeWave')
-            EQN%CCZ4GLMc0   = 1.5   ! 0.1      
-			EQN%CCZ4GLMc    = 1.2   ! 2.0    
-			EQN%CCZ4GLMd    = 2.0   ! 1.0     
-			EQN%CCZ4GLMepsA = 1.0   ! 5. 
-			EQN%CCZ4GLMepsP = 1.0   ! 5.  
-			EQN%CCZ4GLMepsD = 1.0   ! 0.1 
+            EQN%CCZ4GLMc0   = 1.5D0   ! 0.1      
+			EQN%CCZ4GLMc    = 1.2D0   ! 2.0    
+			EQN%CCZ4GLMd    = 2.0D0   ! 1.0     
+			EQN%CCZ4GLMepsA = 1.0D0   ! 5. 
+			EQN%CCZ4GLMepsP = 1.0D0   ! 5.  
+			EQN%CCZ4GLMepsD = 1.0D0   ! 0.1 
 			!
-			EQN%CCZ4itau  = 1.0 
+			EQN%CCZ4itau  = 1.0D0 
 			
-			EQN%CCZ4k1  = 0.0  !modified according to the version in ExaHyPE 1
-			EQN%CCZ4k2  = 0.0 
-			EQN%CCZ4k3  = 0.0 
-			EQN%CCZ4eta = 0.0 
-			EQN%CCZ4f   = 0.0 
-			EQN%CCZ4g   = 0.0 
-			EQN%CCZ4xi  = 0.0 
-			EQN%CCZ4e   = 1.0 
-			EQN%CCZ4c   = 1.0 
-			EQN%CCZ4mu  = 0.2 
-			EQN%CCZ4ds  = 1.0 
-			EQN%CCZ4sk  = 0.0
-			EQN%CCZ4bs   = 0.0      ! set bs=1 if you want to activate the shift convection for beta, b and B (standard CCZ4 formulation). set it to bs=0 to switch off shift convection for those quantities 
+			EQN%CCZ4k1  = 0.0D0  !modified according to the version in ExaHyPE 1
+			EQN%CCZ4k2  = 0.0D0 
+			EQN%CCZ4k3  = 0.0D0 
+			EQN%CCZ4eta = 0.0D0 
+			EQN%CCZ4f   = 0.0D0 
+			EQN%CCZ4g   = 0.0D0 
+			EQN%CCZ4xi  = 0.0D0 
+			EQN%CCZ4e   = 1.0D0 
+			EQN%CCZ4c   = 1.0D0 
+			EQN%CCZ4mu  = 0.2D0 
+			EQN%CCZ4ds  = 1.0D0 
+			EQN%CCZ4sk  = 0.0D0
+			EQN%CCZ4bs   = 0.0D0      ! set bs=1 if you want to activate the shift convection for beta, b and B (standard CCZ4 formulation). set it to bs=0 to switch off shift convection for those quantities 
 			EQN%CCZ4LapseType   = 0 ! harmonic lapse 
-			EQN%EinsteinAutoAux = 0 
+			EQN%EinsteinAutoAux = 0D0 
 
 		case('CCZ4MinkowskiSrc')
 			EQN%CCZ4GLMc0   = 1.5   ! 0.1
